+++ sky/backends/backend_utils.py
@@ -1,5 +1,4 @@
 """Util constants/functions for the backends."""
-import dataclasses
 from datetime import datetime
 import enum
 import getpass
@@ -39,7 +38,6 @@ from sky import provision as provision_lib
 from sky import serve as serve_lib
 from sky import sky_logging
 from sky import skypilot_config
-from sky import spot as spot_lib
 from sky import status_lib
 from sky.backends import onprem_utils
 from sky.provision import instance_setup
@@ -48,6 +46,7 @@ from sky.skylet import log_lib
 from sky.usage import usage_lib
 from sky.utils import command_runner
 from sky.utils import common_utils
+from sky.utils import controller_utils
 from sky.utils import env_options
 from sky.utils import rich_utils
 from sky.utils import subprocess_utils
@@ -151,92 +150,6 @@ _RAY_YAML_KEYS_TO_RESTORE_EXCEPTIONS = [
     ('available_node_types', 'ray.head.default', 'node_config', 'UserData'),
     ('available_node_types', 'ray.worker.default', 'node_config', 'UserData'),
 ]
-
-
-# TODO(tian): Refactor to controller_utils. Current blocker: circular import.
-@dataclasses.dataclass
-class _ControllerSpec:
-    """Spec for skypilot controllers."""
-    name: str
-    cluster_name: str
-    in_progress_hint: str
-    decline_cancel_hint: str
-    decline_down_in_init_status_hint: str
-    decline_down_for_dirty_controller_hint: str
-    check_cluster_name_hint: str
-    default_hint_if_non_existent: str
-
-
-class Controllers(enum.Enum):
-    """Skypilot controllers."""
-    # NOTE(dev): Keep this align with
-    # sky/cli.py::_CONTROLLER_TO_HINT_OR_RAISE
-    SPOT_CONTROLLER = _ControllerSpec(
-        name='managed spot controller',
-        cluster_name=spot_lib.SPOT_CONTROLLER_NAME,
-        in_progress_hint=(
-            '* {job_info}To see all spot jobs: '
-            f'{colorama.Style.BRIGHT}sky spot queue{colorama.Style.RESET_ALL}'),
-        decline_cancel_hint=(
-            'Cancelling the spot controller\'s jobs is not allowed.\nTo cancel '
-            f'spot jobs, use: {colorama.Style.BRIGHT}sky spot cancel <spot '
-            f'job IDs> [--all]{colorama.Style.RESET_ALL}'),
-        decline_down_in_init_status_hint=(
-            f'{colorama.Fore.RED}Tearing down the spot controller while '
-            'it is in INIT state is not supported (this means a spot launch '
-            'is in progress or the previous launch failed), as we cannot '
-            'guarantee that all the spot jobs are finished. Please wait '
-            'until the spot controller is UP or fix it with '
-            f'{colorama.Style.BRIGHT}sky start '
-            f'{spot_lib.SPOT_CONTROLLER_NAME}{colorama.Style.RESET_ALL}.'),
-        decline_down_for_dirty_controller_hint=(
-            f'{colorama.Fore.RED}In-progress spot jobs found. To avoid '
-            f'resource leakage, cancel all jobs first: {colorama.Style.BRIGHT}'
-            f'sky spot cancel -a{colorama.Style.RESET_ALL}\n'),
-        check_cluster_name_hint=(
-            f'Cluster {spot_lib.SPOT_CONTROLLER_NAME} is reserved for '
-            'managed spot controller. '),
-        default_hint_if_non_existent='No managed spot jobs are found.')
-    SKY_SERVE_CONTROLLER = _ControllerSpec(
-        name='sky serve controller',
-        cluster_name=serve_lib.SKY_SERVE_CONTROLLER_NAME,
-        in_progress_hint=(
-            f'* To see detailed service status: {colorama.Style.BRIGHT}'
-            f'sky serve status -a{colorama.Style.RESET_ALL}'),
-        decline_cancel_hint=(
-            'Cancelling the sky serve controller\'s jobs is not allowed.'),
-        decline_down_in_init_status_hint=(
-            f'{colorama.Fore.RED}Tearing down the sky serve controller '
-            'while it is in INIT state is not supported (this means a sky '
-            'serve up is in progress or the previous launch failed), as we '
-            'cannot guarantee that all the services are terminated. Please '
-            'wait until the sky serve controller is UP or fix it with '
-            f'{colorama.Style.BRIGHT}sky start '
-            f'{serve_lib.SKY_SERVE_CONTROLLER_NAME}'
-            f'{colorama.Style.RESET_ALL}.'),
-        decline_down_for_dirty_controller_hint=(
-            f'{colorama.Fore.RED}Tearing down the sky serve controller is not '
-            'supported, as it is currently serving the following services: '
-            '{service_names}. Please terminate the services first with '
-            f'{colorama.Style.BRIGHT}sky serve down -a'
-            f'{colorama.Style.RESET_ALL}.'),
-        check_cluster_name_hint=(
-            f'Cluster {serve_lib.SKY_SERVE_CONTROLLER_NAME} is reserved for '
-            'sky serve controller. '),
-        default_hint_if_non_existent='No service is found.')
-
-    @classmethod
-    def from_name(cls, name: Optional[str]) -> Optional['Controllers']:
-        """Check if the cluster name is a controller name.
-
-        Returns:
-            The controller if the cluster name is a controller name.
-            Otherwise, returns None.
-        """
-        for controller in cls:
-            if controller.value.cluster_name == name:
-                return controller
-        return None
 
 
 def is_ip(s: str) -> bool:
@@ -2589,7 +2502,7 @@ def check_cluster_available(
 
 # TODO(tian): Refactor to controller_utils. Current blocker: circular import.
 def is_controller_up(
-    controller_type: Controllers,
+    controller_type: controller_utils.Controllers,
     stopped_message: str,
     non_existent_message: Optional[str] = None,
 ) -> Tuple[Optional[status_lib.ClusterStatus],
@@ -2702,7 +2615,7 @@ def get_clusters(
     if not include_controller:
         records = [
             record for record in records
-            if Controllers.from_name(record['name']) is None
+            if controller_utils.Controllers.from_name(record['name']) is None
         ]
 
     yellow = colorama.Fore.YELLOW
@@ -2889,29 +2802,6 @@ def get_task_resources_str(task: 'task_lib.Task') -> str:
     resources_str = ', '.join(f'{k}:{v}' for k, v in resources_dict.items())
     resources_str = f'{task.num_nodes}x [{resources_str}]'
     return resources_str
-
-
-# TODO(tian): Refactor to controller_utils. Current blocker: circular import.
-def check_cluster_name_not_controller(
-        cluster_name: Optional[str],
-        operation_str: Optional[str] = None) -> None:
-    """Errors out if the cluster name is a controller name.
-
-    Raises:
-      sky.exceptions.NotSupportedError: if the cluster name is a controller
-        name, raise with an error message explaining 'operation_str' is not
-        allowed.
-
-    Returns:
-      None, if the cluster name is not a controller name.
-    """
-    controller = Controllers.from_name(cluster_name)
-    if controller is not None:
-        msg = controller.value.check_cluster_name_hint
-        if operation_str is not None:
-            msg += f' {operation_str} is not allowed.'
-        with ux_utils.print_exception_no_traceback():
-            raise exceptions.NotSupportedError(msg)
 
 
 # Handle ctrl-c
