+++ sky/utils/controller_utils.py
@@ -1,11 +1,18 @@
 """Util constants/functions for SkyPilot Controllers."""
+import dataclasses
+import enum
 import os
 import typing
 from typing import Optional
 
+import colorama
+
 from sky import exceptions
+from sky import serve as serve_lib
 from sky import sky_logging
+from sky import spot as spot_lib
 from sky.utils import common_utils
+from sky.utils import ux_utils
 
 if typing.TYPE_CHECKING:
     from sky.backends import cloud_vm_ray_backend
@@ -15,6 +22,113 @@ logger = sky_logging.init_logger(__name__)
 # The default idle timeout for skypilot controllers. This include spot
 # controller and sky serve controller.
 CONTROLLER_IDLE_MINUTES_TO_AUTOSTOP = 10
+
+
+@dataclasses.dataclass
+class _ControllerSpec:
+    """Spec for skypilot controllers."""
+    name: str
+    cluster_name: str
+    in_progress_hint: str
+    decline_cancel_hint: str
+    decline_down_in_init_status_hint: str
+    decline_down_for_dirty_controller_hint: str
+    check_cluster_name_hint: str
+    default_hint_if_non_existent: str
+
+
+class Controllers(enum.Enum):
+    """Skypilot controllers."""
+    # NOTE(dev): Keep this align with
+    # sky/cli.py::_CONTROLLER_TO_HINT_OR_RAISE
+    SPOT_CONTROLLER = _ControllerSpec(
+        name='managed spot controller',
+        cluster_name=spot_lib.SPOT_CONTROLLER_NAME,
+        in_progress_hint=(
+            '* {job_info}To see all spot jobs: '
+            f'{colorama.Style.BRIGHT}sky spot queue{colorama.Style.RESET_ALL}'),
+        decline_cancel_hint=(
+            'Cancelling the spot controller\'s jobs is not allowed.\nTo cancel '
+            f'spot jobs, use: {colorama.Style.BRIGHT}sky spot cancel <spot '
+            f'job IDs> [--all]{colorama.Style.RESET_ALL}'),
+        decline_down_in_init_status_hint=(
+            f'{colorama.Fore.RED}Tearing down the spot controller while '
+            'it is in INIT state is not supported (this means a spot launch '
+            'is in progress or the previous launch failed), as we cannot '
+            'guarantee that all the spot jobs are finished. Please wait '
+            'until the spot controller is UP or fix it with '
+            f'{colorama.Style.BRIGHT}sky start '
+            f'{spot_lib.SPOT_CONTROLLER_NAME}{colorama.Style.RESET_ALL}.'),
+        decline_down_for_dirty_controller_hint=(
+            f'{colorama.Fore.RED}In-progress spot jobs found. To avoid '
+            f'resource leakage, cancel all jobs first: {colorama.Style.BRIGHT}'
+            f'sky spot cancel -a{colorama.Style.RESET_ALL}\n'),
+        check_cluster_name_hint=(
+            f'Cluster {spot_lib.SPOT_CONTROLLER_NAME} is reserved for '
+            'managed spot controller. '),
+        default_hint_if_non_existent='No managed spot jobs are found.')
+    SKY_SERVE_CONTROLLER = _ControllerSpec(
+        name='sky serve controller',
+        cluster_name=serve_lib.SKY_SERVE_CONTROLLER_NAME,
+        in_progress_hint=(
+            f'* To see detailed service status: {colorama.Style.BRIGHT}'
+            f'sky serve status -a{colorama.Style.RESET_ALL}'),
+        decline_cancel_hint=(
+            'Cancelling the sky serve controller\'s jobs is not allowed.'),
+        decline_down_in_init_status_hint=(
+            f'{colorama.Fore.RED}Tearing down the sky serve controller '
+            'while it is in INIT state is not supported (this means a sky '
+            'serve up is in progress or the previous launch failed), as we '
+            'cannot guarantee that all the services are terminated. Please '
+            'wait until the sky serve controller is UP or fix it with '
+            f'{colorama.Style.BRIGHT}sky start '
+            f'{serve_lib.SKY_SERVE_CONTROLLER_NAME}'
+            f'{colorama.Style.RESET_ALL}.'),
+        decline_down_for_dirty_controller_hint=(
+            f'{colorama.Fore.RED}Tearing down the sky serve controller is not '
+            'supported, as it is currently serving the following services: '
+            '{service_names}. Please terminate the services first with '
+            f'{colorama.Style.BRIGHT}sky serve down -a'
+            f'{colorama.Style.RESET_ALL}.'),
+        check_cluster_name_hint=(
+            f'Cluster {serve_lib.SKY_SERVE_CONTROLLER_NAME} is reserved for '
+            'sky serve controller. '),
+        default_hint_if_non_existent='No service is found.')
+
+    @classmethod
+    def from_name(cls, name: Optional[str]) -> Optional['Controllers']:
+        """Check if the cluster name is a controller name.
+
+        Returns:
+            The controller if the cluster name is a controller name.
+            Otherwise, returns None.
+        """
+        for controller in cls:
+            if controller.value.cluster_name == name:
+                return controller
+        return None
+
+
+def check_cluster_name_not_controller(
+        cluster_name: Optional[str],
+        operation_str: Optional[str] = None) -> None:
+    """Errors out if the cluster name is a controller name.
+
+    Raises:
+      sky.exceptions.NotSupportedError: if the cluster name is a controller
+        name, raise with an error message explaining 'operation_str' is not
+        allowed.
+
+    Returns:
+      None, if the cluster name is not a controller name.
+    """
+    controller = Controllers.from_name(cluster_name)
+    if controller is not None:
+        msg = controller.value.check_cluster_name_hint
+        if operation_str is not None:
+            msg += f' {operation_str} is not allowed.'
+        with ux_utils.print_exception_no_traceback():
+            raise exceptions.NotSupportedError(msg)
 
 
 # Internal only:
